add commands/setting.ts
