+++ src/lib/useSettings.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { Settings } from './types';
-import { getSettings } from '@/commands/account';
+import { getSettings } from '@/commands/setting';
 
 export function useSettings() {
   return useQuery<Settings, Error>({
